Move StateSnapshot into its own file.
